# internal/api/handlers/user/bind_discord.go
package user

import (
	"net/http"
	"strconv"

	api_utils "github.com/dzhisl/license-api/internal/api/utils"
	"github.com/dzhisl/license-api/internal/storage"
	"github.com/dzhisl/license-api/pkg/logger"
	"github.com/gin-gonic/gin"
	"go.uber.org/zap"
)

type bindDiscordRequest struct {
	DiscordId int `json:"discord_id" binding:"required"`
}

// @Summary Bind Discord to user
// @Description Associates a Discord ID with the user specified by user_id in the URL.
// @Tags user
// @Accept json
// @Produce json
// @Param user_id path int true "User ID"
// @Param request body bindDiscordRequest true "payload"
// @Success 200 {object} statusResponse
// @Failure 400 {object} invalidBodyErrResponse
// @Failure 500 {object} internalErrResponse
// @Security ApiKeyAuth
// @Router /user/{user_id}/discord [post]
func BindDiscordHandler(c *gin.Context) {
	conn := storage.GetConnector()
	ctx := c.Request.Context()

	userIdStr := c.Param("user_id")
	userId, err := strconv.Atoi(userIdStr)
	if err != nil {
		logger.Debug(ctx, "invalid user_id", zap.Error(err))
		c.JSON(api_utils.FormErrResponse(http.StatusBadRequest, "user_id must be an integer"))
		return
	}

	var req bindDiscordRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		logger.Debug(ctx, "invalid request body", zap.Error(err))
		c.JSON(api_utils.FormInvalidRequestResponse())
		return
	}

	err = conn.BindDiscord(ctx, userId, req.DiscordId)
	if err != nil {
		logger.Error(ctx, "failed to bind discord", zap.Error(err))
		c.JSON(api_utils.FormInternalErrResponse())
		return
	}

	c.JSON(http.StatusOK, statusResponse{Status: "success"})
}